% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\docType{data}
\name{municipality}
\alias{municipality}
\title{New Jersey municipality names and GEOIDs from the US Census}
\format{
A dataframe with 3 columns
\describe{
 \item{county}{County name, e.g. "Atlantic County"}
 \item{municipality}{Municipalilty name, e.g. "Buena borough"}
 \item{GEOID}{US Census Geographic Identifier}
}
}
\usage{
municipality
}
\description{
New Jersey municipality names and GEOIDs from the US Census
}
\keyword{datasets}
